Move decryptFile into a utility function so that it can be shared between different components
